+++ app/components/bulma/tag_component.rb
@@ -2,14 +2,15 @@
 
 module Bulma
   class TagComponent < Component
-    def initialize(color: nil, size: :normal, light: false)
+    def initialize(label = nil, color: nil, size: :normal, light: false)
+      @label = label
       @size = size
       @color = color
       @light = light
     end
 
     def call
-      content_tag :span, content, class: class_names(
+      content_tag :span, @label || content, class: class_names(
         "tag",
         "is-light" => @light,
         "is-#{@color}" => @color,