+++ previews/elements/tag_component_preview.rb
@@ -2,7 +2,6 @@
 
 class Elements::TagComponentPreview < ViewComponent::Preview
   def default
-    render(Bulma::TagComponent.new) { "Tag label" }
   end
 
   def colors